Add files via upload
This is a mock-up of the display / MMI. Currently has a rather obvious flaw in it - all parameters need to be set each time one is changed. :-( Partly a product of using a Variable Resistor and A/D converter for the input. Plan to change this to a rotary encoder and this will remove this obvious defect.