Add files via upload
Working - but untested version of the VC Controller for Arduino. Excel file is used to build the drive 'waveform' .. i.e. to shape the rate at which air driven to the patient during an inhillation cycle. Excel generates a data array that is then copied into the Arduino code. This versio of the code includes (1) Simple, 'single switch and rotation' interface for Respitory rate, I:E ratio and Tidal volume (2) Drive waveform as shaped by the definition array (generated from Excel at design time); (3) The three control parameters then used to drive the output. Currently the 'output' is only a print to the serial interface for testing. [TO DO] (1) Driver loop to run from an interrupt so that it does not halt whilst changing parameters (2) Switch output to PWM to drive the real hardware.