Add files via upload
This version runs the main vent control loop via an interrupt timer. This means that control is not paused whilst parameters are changed. TODO: (1) The parameters should only be swapped into the control loop at the end of a stroke - not arbitrarily as they currently are. (2) There may be race conditions with updating control variables and these being read inside the main control loop. Interrupts should be paused momentarily during the final update of control parameter (i.e. at the moment that the new control parameters are 'swapped in'