Add files via upload
Fixes in this version of the code (1) Changed some badly named constant definitions (2) Controller only reacts to parameter changes at the end of a breath cycle .. between selecting a new parameter value and the start of the next breath cycle the message "Adjusting .. " is displayed (3) Introduced a 'semaphore' between the human-interace code to enter parameters and the use of those parameters within the (timer interrupt driver) main control loop. This prevents a race condition and removes the opportunity to read partially-written variables within the main control loop.

The PDF file is the corresponding Arduino bread-board circuit.