Add files via upload
I added a 'toy' servo - to provide a physical indication of what the controller is doing. Here is the code - and the circuit. I plan to use this as my simple 'bench-test' setup. Also, I would like to get some of these in the hands of those with medical knowledge on the team to confirm the human interface. I believe this does what I was asked for .. but the I.E. ratio behaviour does not really make sense to me.