Add files via upload
This version includes a 'calibration mode' (for test and development purposes only). When the select key is held down for 5 seconds, the controller enters 'calibration mode'. This enables the test engineer to enter 'raw' values to be sent to the actuator. From this we can calibrate actator input to angular rotation of arm, and to 'volume of gas' squeezed from the bag.