Add files via upload
Increased maximum tidal volume from 250 to 700 ml. Note that this is only a scalling factor on the digital (PWM) output. That output then needs to be calibrated and transformed to match the actual volume of gas pumped by the device mechanics. That transformation beteen internal values and external (physical) volume is not included in this version and is a 'TODO'